requirements.txt

langchain==0.0.148

llama-index==0.5.6


python-decouple==3.8
fastapi
uvicorn
python-dotenv 
openai